Address some pr comments - remove ABC_EQF.cpp, clean up comments, change Input class to struct, remove changes to target
