+++ gtsam/3rdparty/metis/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.5...3.22)
+cmake_minimum_required(VERSION 3.5)
 project(METIS)
 
 # Add flags for currect directory and below
